requirements/base.txt
django==3.1.2
django-filter==2.4.0
graphene-django==2.13.0
django-graphql-jwt==0.3.0
PyJWT==1.7.1
colorama==0.4.4
python-dotenv==0.15.0
django-mail-templated==2.6.5
twilio==6.61.0
phonenumbers==8.12.28